pyproject.toml
[tool.poetry]
name = "stable-diffusion-botui"
version = "0.1.0"
description = ""
authors = ["daima3629"]
license = "MIT"
readme = "README.md"
packages = [{include = "stable_diffusion_botui"}]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.1.0"
aiohttp = "^3.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
